pyproject.toml
[tool.poetry]
name = "kafka-uptime-monitor"
version = "0.1.0"
description = ""
authors = ["Jussi Tikkanen <drjustigious@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.17.1"
requests = "^2.25.1"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
